vld尚可用。

+++ StockAnalysis/source/Data Source/NeteaseRTDataSource.cpp
@@ -18,8 +18,8 @@ bool CNeteaseRTDataSource::UpdateStatus(void) {
 }
 
 bool CNeteaseRTDataSource::Inquire(long lCurrentTime) {
-	static long long sllLastTimeTickCount = 0;
-	long long llTickCount = 0;
+	static ULONGLONG sllLastTimeTickCount = 0;
+	ULONGLONG llTickCount = 0;
 
 	llTickCount = GetTickCount64();
 	if (llTickCount > (sllLastTimeTickCount + gl_systemConfiguration.GetChinaMarketRTDataInquiryTime())) {
@@ -27,9 +27,7 @@ bool CNeteaseRTDataSource::Inquire(long lCurrentTime) {
 			InquireRTData(lCurrentTime);
 		}
 		if (IsInquiring()) {
-			if (!gl_pChinaMarket->IsFastReceivingRTData() && gl_pChinaMarket->IsSystemReady() && !gl_systemConfiguration.
-				IsDebugMode()) {
-				// 系统配置为：测试系统时，不降低轮询速度
+			if (!gl_pChinaMarket->IsFastReceivingRTData() && gl_pChinaMarket->IsSystemReady() && !gl_systemConfiguration.IsDebugMode()) {	// 系统配置为：测试系统时，不降低轮询速度
 				sllLastTimeTickCount = llTickCount + 60000; // 完全轮询一遍后，非交易时段一分钟左右更新一次即可
 			}
 			else {
